c11/c11_thread01/test.c: Address printf format specifier warning
The type thrd_t is opaque and does not have a format specifier.
Knowing it is an RTEMS id, changed to cast it to unsigned long

+++ c11/c11_thread01/test.c
@@ -35,7 +35,7 @@ int Thread1_Body(void *arg)
 
   initialize_subsystem();
 
-  printf("Thread1(0x%08lx)\n", *t);
+  printf("Thread1(0x%08lx)\n", (unsigned long) *t);
   assert(thrd_current() == thread1);
 
   puts("Thread1 - yield" );
@@ -57,7 +57,7 @@ int Thread2_Body(void *arg)
 
   initialize_subsystem();
 
-  printf("Thread2(0x%08lx)\n", *t);
+  printf("Thread2(0x%08lx)\n", (unsigned long) *t);
   assert(thrd_current() == thread2);
 
   puts("Thread2 - exit" );
